Исправлены баги, добавлены DTO для POST методов

+++ src/main/java/com/github/polliakov/bank/services/impl/CreditServiceImpl.java
@@ -1,26 +1,44 @@
 package com.github.polliakov.bank.services.impl;
 
+import com.github.polliakov.bank.dto.CreditDto;
 import com.github.polliakov.bank.entities.BankEntity;
 import com.github.polliakov.bank.entities.CreditEntity;
+import com.github.polliakov.bank.entities.CreditOfferEntity;
+import com.github.polliakov.bank.repositories.BankRepository;
 import com.github.polliakov.bank.repositories.CreditRepository;
 import com.github.polliakov.bank.services.CreditService;
+import com.github.polliakov.bank.services.DtoMapperService;
 import org.springframework.stereotype.Service;
 
 import java.lang.module.FindException;
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 @Service
 public class CreditServiceImpl implements CreditService {
-    public CreditServiceImpl(CreditRepository repository) {
+    public CreditServiceImpl(BankRepository bankRepository,
+                             CreditRepository repository,
+                             DtoMapperService dtoMapper) {
+        this.bankRepository = bankRepository;
         this.repository = repository;
+        this.dtoMapper = dtoMapper;
     }
 
+    private final BankRepository bankRepository;
     private final CreditRepository repository;
+    private final DtoMapperService dtoMapper;
 
     @Override
-    public void create(CreditEntity creditEntity) {
-        creditEntity.setId(null);
-        repository.save(creditEntity);
+    public void create(CreditDto creditDto) {
+        var credit = dtoMapper.entityFromDto(creditDto);
+        credit.setId(null);
+        repository.save(credit);
+
+        var banks = credit.getBanks();
+        if (banks == null) return;
+        for (var b : banks)
+            b.getCredits().add(credit);
+        bankRepository.saveAll(banks);
     }
 
     @Override
@@ -34,11 +52,19 @@ public class CreditServiceImpl implements CreditService {
     }
 
     @Override
-    public List<BankEntity> getBanks(Long creditId){
+    public List<BankEntity> getBanks(Long creditId) {
         var credit = getById(creditId);
         if (credit == null)
             throw new FindException("Credit with id = " + creditId + "is not found");
-        return new ArrayList<>(credit.getBanks());
+        return new LinkedList<>(credit.getBanks());
+    }
+
+    @Override
+    public List<CreditOfferEntity> getCreditOffers(Long creditId) {
+        var credit = getById(creditId);
+        if (credit == null)
+            throw new FindException("Credit with id = " + creditId + "is not found");
+        return new LinkedList<>(credit.getOfferEntities());
     }
 
     @Override
@@ -49,8 +75,11 @@ public class CreditServiceImpl implements CreditService {
 
     @Override
     public void deleteById(Long id) {
-        CheckExists(id);
-        repository.deleteById(id);
+        var credit = repository.findById(id).orElseThrow();
+        if (credit.getOfferEntities().isEmpty())
+            repository.deleteById(id);
+        else
+            throw new UnsupportedOperationException();
     }
 
     private void CheckExists(Long id) {
